- Renamed NewConstantPropagation.java to ConstantPropagation - Added new Option to allow propagation of copy assignments
