+++ dartagnan/src/main/java/com/dat3m/dartagnan/configuration/OptionNames.java
@@ -35,6 +35,7 @@ public class OptionNames {
 	public static final String CONSTANT_PROPAGATION = "program.processing.constantPropagation";
 	public static final String DEAD_ASSIGNMENT_ELIMINATION = "program.processing.dce";
 	public static final String DYNAMIC_PURE_LOOP_CUTTING = "program.processing.dplc";
+	public static final String PROPAGATE_COPY_ASSIGNMENTS = "program.processing.propagateCopyAssignments";
 	
 	// Program Property Options
 	public static final String ALIAS_METHOD = "program.analysis.alias";
